Trait plasma::Mixer

pub trait Mixer<T: Sized + Default + Copy> {
    type IntermediateH: Sized + Default + Copy + BorrowMut<[T]> + Borrow<[T]>;
    type IntermediateV: Sized + Default + Copy + BorrowMut<[T]> + Borrow<[T]>;
    fn mix_pixels(
        &self, 
        vxp: &Self::IntermediateH, 
        vyp: &Self::IntermediateV, 
        next_pixel: &mut dyn FnMut(PixelRgb)
    );

    fn intermediate_h_len() -> usize { ... }

    fn intermediate_v_len() -> usize { ... }
}

Implementations of this trait should compute the color of each pixel based on an intermediate data created by a IntermediateCalculator.

The type T should be a f32 or a packed simd f32x8 if a "use-simd" crate feature is enabled.

Associated Types

type IntermediateH: Sized + Default + Copy + BorrowMut<[T]> + Borrow<[T]>

This type should be an array of the type T for an intermediate horizontal data.

type IntermediateV: Sized + Default + Copy + BorrowMut<[T]> + Borrow<[T]>

This type should be an array of the type T for an intermediate vertical data.

Required Methods

fn mix_pixels(
&self,
vxp: &Self::IntermediateH,
vyp: &Self::IntermediateV,
next_pixel: &mut dyn FnMut(PixelRgb)
)

The implementors should compute a pixel and send it as an instance of PixelRgb to the provided next_pixel function.

The computation should be based on the provided combination of intermediate data.

Provided Methods

fn intermediate_h_len() -> usize

Returns the number of intermediate horizontal values.

fn intermediate_v_len() -> usize

Returns the number of intermediate vertical values.